requirements.txt
# requirements.txt

click==8.1.8
customtkinter==5.2.2
darkdetect==0.8.0
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
mysql-connector==2.2.9
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.1.0
setuptools==75.6.0
wheel==0.45.1
